# shared/contracts/openapi.tenant.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TexQtic Tenant API",
    "version": "0.3.0",
    "description": "Tenant-facing API surface for TexQtic multi-tenant SaaS platform. Includes auth, catalog, cart, AI, and tenant management."
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "tenantJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Tenant JWT token obtained from /api/auth/tenant/login or /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/tenant/certifications": {
      "post": {
        "summary": "Create a certification record",
        "tags": ["Tenant - Certifications"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["certificationType", "reason"],
                "properties": {
                  "certificationType": { "type": "string" },
                  "reason": { "type": "string" },
                  "issuedAt": { "type": "string", "format": "date-time" },
                  "expiresAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Certification created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificationId": { "type": "string", "format": "uuid" },
                    "stateKey": { "type": "string" },
                    "certificationType": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "INVALID_INPUT" },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN" },
          "500": {
            "description": "DB_ERROR — server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List certifications for authenticated tenant",
        "tags": ["Tenant - Certifications"],
        "parameters": [
          { "in": "query", "name": "stateKey", "schema": { "type": "string" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer" } },
          { "in": "query", "name": "offset", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of certifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "type": "object" } },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN" }
        }
      }
    },
    "/api/tenant/certifications/{id}": {
      "get": {
        "summary": "Get certification detail",
        "tags": ["Tenant - Certifications"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Certification detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certification": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN" },
          "404": { "description": "NOT_FOUND" }
        }
      },
      "patch": {
        "summary": "Update certification metadata",
        "tags": ["Tenant - Certifications"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "certificationType": { "type": "string" },
                  "issuedAt": { "type": "string", "format": "date-time" },
                  "expiresAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Certification updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "400": { "description": "INVALID_INPUT" },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN" },
          "404": { "description": "NOT_FOUND" }
        }
      }
    },
    "/api/tenant/certifications/{id}/document/upload": {
      "post": {
        "summary": "Upload or replace a private certificate document",
        "tags": ["Tenant - Certifications"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "PDF, JPG, PNG, or WEBP certificate document. Max 5 MB."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Certificate document metadata saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificationId": { "type": "string", "format": "uuid" },
                    "documentOriginalName": {
                      "type": "string",
                      "nullable": true
                    },
                    "documentMimeType": { "type": "string" },
                    "documentSizeBytes": { "type": "integer" },
                    "documentUploadedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "FILE_REQUIRED, FILE_TOO_LARGE, or INVALID_FILE_TYPE"
          },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN — OWNER or ADMIN required" },
          "404": { "description": "NOT_FOUND" },
          "500": { "description": "STORAGE_NOT_CONFIGURED or UPLOAD_FAILED" }
        }
      }
    },
    "/api/tenant/certifications/{id}/document": {
      "get": {
        "summary": "Get authenticated short-lived access to a private certificate document",
        "tags": ["Tenant - Certifications"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Short-lived signed certificate document access URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificationId": { "type": "string", "format": "uuid" },
                    "signedUrl": { "type": "string", "format": "uri" },
                    "documentOriginalName": {
                      "type": "string",
                      "nullable": true
                    },
                    "documentMimeType": { "type": "string", "nullable": true },
                    "documentSizeBytes": {
                      "type": "integer",
                      "nullable": true
                    },
                    "documentUploadedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "UNAUTHORIZED" },
          "403": { "description": "FORBIDDEN" },
          "404": { "description": "NOT_FOUND" },
          "500": {
            "description": "STORAGE_NOT_CONFIGURED or SIGNED_URL_FAILED"
          }
        }
      }
    },
...
  }
}